Removes caching from WaterbalanceComputer2.get_load_timeseries
Subsequent calls to this method always return the same objects thanks to to
built-in caching. Unfortunately, the built-in caching broke due to the
addditional method parameter added by commit

   1a651a0fe8ff5b9df8872fa02ba4deeeb96dc087

and subsequent calls should not return the same objects anymore.

This commit removes the caching as it should not have been build in in the
first place.