Sanitizes the output of LoadComputer.compute, #38
Before this commit, method LoadComputer.compute returned a pair of
dictionaries, where each dictionary mapped a string or a PumpingStation to
a time series. Consequently, functions that used these dictionaries had to
inspect the key and its type before they could use the time series. which
lead to a lot of code duplication.

To overcome that problem, this commit modifies LoadComputer.compute to
return a list of Load objects.