Fixes the naming of the undrained concentration, #38
The Bucket attributes that specifies the minimum and incremental concentrations
of the 'undrained' bucket flow are named min_concentr_nitrogen_drainage_indraft
and incr_concentr_nitrogen_drainage_indraft. This is unfortunate as the other
concentrations are named according to the bucket flow names.

The exception mentioned above would also cause some lower level selection
statements. To avoid that, we change the internal naming of these concetrations
to min_concentr_nitrogen_undrained and inc_concentr_nitrogen_undrained.