Refactors the code of ConcentrationComputer::compute, #46
The body of the loop of ConcentrationComputer::compute computes the volume, the
chloride level and the concentration at the end of a specific day. To me, that
was not obvious from the implementation so I decided to refactor it. This
commit improves the situation, but some further work might be in order.